Fixed code

+++ src/main/java/lv/onlineStore/demo/repos/iCourseRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import lv.onlineStore.demo.models.Course;
 
-public interface iCourseRepo extends CrudRepository<Course, Long> {
+public interface iCourseRepo extends CrudRepository<Course, Integer> {
 
 }
