+++ src/main/java/lv/onlineStore/demo/repos/iStudentRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import lv.onlineStore.demo.models.Student;
 
-public interface iStudentRepo extends CrudRepository<Student, Long>{
+public interface iStudentRepo extends CrudRepository<Student, Integer>{
 
 }
